[frontend] refactor existing Tags component; add Tags field to the register form
