+++ hwproj.front/src/components/Homeworks/AddHomework.tsx
@@ -6,8 +6,9 @@ import {TextFieldWithPreview} from "../Common/TextFieldWithPreview";
 import {CreateHomeworkViewModel, CreateTaskViewModel} from "../../api";
 import PublicationAndDeadlineDates from "../Common/PublicationAndDeadlineDates";
 import CreateTask from "../Tasks/CreateTask"
-import {Alert, Checkbox, FormControlLabel} from "@mui/material";
-import Tags from "./HomeworkTags";
+import Tags from "../Common/Tags";
+import apiSingleton from "../../api/ApiSingleton";
+import {Alert } from "@mui/material";
 import { TestTag } from "components/Common/HomeworkTags";
 
 interface IAddHomeworkProps {
@@ -122,7 +123,8 @@ const AddHomework: React.FC<IAddHomeworkProps> = (props) => {
                     }
                     }
                 />
-                <Tags tags={[]} courseId={props.id} onTagsChange={handleTagsChange}/>
+                <Tags tags={[]} onTagsChange={handleTagsChange} isAutocompleteSmall={false}
+                      requestTags={() => apiSingleton.coursesApi.apiCoursesTagsByCourseIdGet(props.id)}/>
                 {addHomeworkState.tags.includes(TestTag) && <Alert severity="info">Вы можете сгруппировать контрольные работы и переписывания с помощью дополнительного тега. Например, 'КР 1'</Alert>}
                 <PublicationAndDeadlineDates
                     hasDeadline={false}
